+++ src/public-api.ts
@@ -177,3 +177,8 @@ export * from "./content/vehicle/VehicleMarketDemandContentValidator.js";
 export * from "./content/vehicle/VehicleStage15Validator.js";
 export * from "./application/services/VehicleMarketDemandService.js";
 export * from "./application/vehicle-market/VehicleVariantLifecycleRefresher.js";
+
+export * from "./content/map/WorldMapContent.js";
+export * from "./content/map/WorldMapContentValidator.js";
+export * from "./content/map/FormalWorldMapContent.js";
+export * from "./content/map/WorldMapSeed.js";